New: allowed for nesting child attributes infinitely
