+++ app/apollo/resolvers/subscription.js
@@ -23,6 +23,7 @@ const { whoIs, validAuth, NotFoundError, validClusterAuth, getGroupConditions, g
 const getSubscriptionUrls = require('../../utils/subscriptions.js').getSubscriptionUrls;
 const { EVENTS, GraphqlPubSub, getStreamingTopic } = require('../subscription');
 const GraphqlFields = require('graphql-fields');
+const { applyQueryFieldsToSubscriptions } = require('../utils/applyQueryFields');
 
 const pubSub = GraphqlPubSub.getInstance();
 
@@ -44,75 +45,7 @@ async function validateGroups(org_id, groups, context) {
   return groupCount;
 }
 
-const applyQueryFieldsToSubscriptions = async(subs, queryFields, { orgId }, models)=>{ // eslint-disable-line
-  _.each(subs, (sub)=>{
-    if(_.isUndefined(sub.channelName)){
-      sub.channelName = sub.channel;
-    }
-    delete sub.channel;
-  });
-  var subUuids = _.uniq(_.map(subs, 'uuid'));
 
-  if(queryFields.channel && subs.length > 0){
-    var channelUuids = _.uniq(_.map(subs, 'channelUuid'));
-    var channels = await models.Channel.find({ uuid: { $in: channelUuids } });
-    var channelsByUuid = _.keyBy(channels, 'uuid');
-    _.each(subs, (sub)=>{
-      var channelUuid = sub.channelUuid;
-      var channel = channelsByUuid[channelUuid];
-      sub.channel = channel;
-    });
-  }
-  if(queryFields.resources && subs.length > 0){
-    var resources = await models.Resource.find({ org_id: orgId, 'searchableData.subscription_id' : { $in: subUuids } }).lean({virtuals: true});
-    var resourcesBySubUuid = _.groupBy(resources, 'searchableData.subscription_id');
-    _.each(subs, (sub)=>{
-      sub.resources = resourcesBySubUuid[sub.uuid] || [];
-    });
-  }
-  if(queryFields.groupObjs){
-    var groupNames = _.flatten(_.map(subs, 'groups'));
-    var groups = await models.Group.find({ org_id: orgId, name: { $in: groupNames } });
-    var groupsByName = _.keyBy(groups, 'name');
-    _.each(subs, (sub)=>{
-      sub.groupObjs = _.filter(_.map(sub.groups, (groupName)=>{
-        return groupsByName[groupName];
-      }));
-    });
-  }
-  if(queryFields.remoteResources || queryFields.rolloutStatus){
-    var remoteResources = await models.Resource.find({
-      org_id: orgId,
-      'searchableData.annotations["deploy_razee_io_clustersubscription"]': { $in: subUuids },
-      deleted: false,
-    });
-    var remoteResourcesBySubUuid = _.groupBy(remoteResources, (rr)=>{
-      return rr.searchableData.get('annotations["deploy_razee_io_clustersubscription"]');
-    });
-    _.each(subs, (sub)=>{
-      var rrs = remoteResourcesBySubUuid[sub.uuid] || [];
-
-      // loops through each resource. if there are errors, increments the errorCount. if no errors, increments successfulCount
-      var errorCount = 0;
-      var successCount = 0;
-      _.each(rrs, (rr)=>{
-        var errors = _.toArray(rr.searchableData.get('errors')||[]);
-        if(errors.length > 0){
-          errorCount += 1;
-        }
-        else{
-          successCount += 1;
-        }
-      });
-
-      sub.remoteResources = rrs;
-      sub.rolloutStatus = {
-        successCount,
-        errorCount,
-      };
-    });
-  }
-};
 
 const subscriptionResolvers = {
   Query: {
