+++ app/apollo/utils/applyQueryFields.js
@@ -1,0 +1,265 @@
+/**
+ * Copyright 2020 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+var _ = require('lodash');
+
+const applyQueryFieldsToClusters = async(clusters, queryFields={}, args, models)=>{
+  var { orgId, resourceLimit, groupLimit } = args;
+
+  const clusterIds = _.map(clusters, 'cluster_id');
+  resourceLimit = resourceLimit || 500;
+
+  _.each(clusters, (cluster)=>{
+    cluster.name = cluster.name || cluster.clusterId || cluster.id;
+  });
+  if(queryFields.resources) {
+    const resources = await models.Resource.find({ cluster_id: { $in: clusterIds } }).limit(resourceLimit).lean({virtuals: true});
+    await applyQueryFieldsToResources(resources, queryFields.resources, args, models);
+
+    const resourcesByClusterId = _.groupBy(resources, 'cluster_id');
+    _.each(clusters, (cluster) => {
+      cluster.resources = resourcesByClusterId[cluster.cluster_id] || [];
+    });
+  }
+  if(queryFields.groupObjs){
+    if(clusterIds.length > 0){
+      // [
+      //   {groups: [{name: 'tag1'}]},
+      //   {groups: [{name: 'tag2'}]},
+      // ]
+      var groupNames = _.filter(_.uniq(_.map(_.flatten(_.map(clusters, 'groups')), 'name')));
+      var groups = await models.Group.find({ org_id: orgId, name: { $in: groupNames } }).limit(groupLimit).lean({ virtuals: true });
+      await applyQueryFieldsToGroups(groups, queryFields.groupObjs, args, models);
+
+      var groupsByUuid = _.keyBy(groups, 'uuid');
+      _.each(clusters, (cluster)=>{
+        var clusterGroupUuids = _.map(cluster.groups, 'uuid');
+        cluster.groupObjs = _.filter(_.pick(groupsByUuid, clusterGroupUuids));
+      });
+    }
+  }
+};
+
+const applyQueryFieldsToGroups = async(groups, queryFields={}, args, models)=>{
+  var { orgId } = args;
+
+  if(queryFields.owner){
+    const owners = await models.User.getBasicUsersByIds(_.uniq(_.map(groups, 'owner')));
+    _.each(groups, (group)=>{
+      group.owner = owners[group.owner];
+    });
+  }
+  if(queryFields.subscriptions || queryFields.subscriptionCount){
+    var groupNames = _.uniq(_.map(groups, 'name'));
+    var subscriptions = await models.Subscription.find({ org_id: orgId, groups: { $in: groupNames } }).lean({ virtuals: true });
+    await applyQueryFieldsToSubscriptions(subscriptions, queryFields.subscriptions, args, models);
+
+    // if(queryFields.subscriptions && queryFields.subscriptions.owner && subscriptions) {
+    //   const ownerIds = _.map(subscriptions, 'owner');
+    //   const subOwners = await models.User.getBasicUsersByIds(ownerIds);
+    //   subscriptions = subscriptions.map((sub)=>{
+    //     sub.owner = subOwners[sub.owner];
+    //     return sub;
+    //   });
+    // }
+    const subscriptionsByGroupName = {};
+    _.each(subscriptions, (sub)=>{
+      if(_.isUndefined(sub.channelName)){
+        sub.channelName = sub.channel;
+      }
+      delete sub.channel; // can not render channel, too deep
+      _.each(sub.groups, (groupName)=>{
+        subscriptionsByGroupName[groupName] = subscriptionsByGroupName[groupName] || [];
+        subscriptionsByGroupName[groupName].push(sub);
+      });
+    });
+    _.each(groups, (group)=>{
+      group.subscriptions = subscriptionsByGroupName[group.name] || [];
+      group.subscriptionCount = group.subscriptions.length;
+    });
+    // if(_.get(queryFields, 'subscriptions.resources')){
+    //   var subUuids = _.uniq(_.map(subscriptions, 'uuid'));
+    //   var resources = await models.Resource.find({ org_id: orgId, 'searchableData.subscription_id': { $in: subUuids }});
+    //   var resourcesBySubUuid = _.groupBy(resources, (resource)=>{
+    //     return resource.searchableData.get('subscription_id');
+    //   });
+    //   _.each(groups, (group)=>{
+    //     _.each(group.subscriptions, (sub)=>{
+    //       sub.resources = resourcesBySubUuid[sub.uuid] || [];
+    //     });
+    //   });
+    // }
+  }
+  if(queryFields.clusters || queryFields.clusterCount){
+    var groupUuids = _.uniq(_.map(groups, 'uuid'));
+    const clusters = await models.Cluster.find({ org_id: orgId, 'groups.uuid': { $in: groupUuids } }).lean({ virtuals: true });
+    await applyQueryFieldsToClusters(clusters, queryFields.clusters, args, models);
+
+    const clustersByGroupUuid = {};
+    _.each(clusters, (cluster)=>{
+      _.each(cluster.groups || [], (groupObj)=>{
+        clustersByGroupUuid[groupObj.uuid] = clustersByGroupUuid[groupObj.uuid] || [];
+        clustersByGroupUuid[groupObj.uuid].push(cluster);
+      });
+    });
+    _.each(groups, (group)=>{
+      group.clusters = clustersByGroupUuid[group.uuid] || [];
+      group.clusterCount = group.clusters.length;
+    });
+  }
+};
+
+const applyQueryFieldsToResources = async(resources, queryFields={}, args, models)=>{
+  var { orgId, subscriptionsLimit = 500 } = args;
+
+  if(queryFields.cluster){
+    var clusterIds = _.map(resources, 'clusterId');
+    var clusters = await models.Cluster.find({ org_id: orgId, cluster_id: { $in: clusterIds } }).lean({ virtuals: true });
+    await applyQueryFieldsToClusters(clusters, queryFields.cluster, args, models);
+
+    var clustersById = _.keyBy(clusters, 'clusterId');
+    _.each(resources, (resource)=>{
+      resource.cluster = clustersById[resource.clusterId];
+    });
+  }
+
+  if(queryFields.subscription){
+    var subscriptionUuids = _.filter(_.uniq(_.map(resources, 'searchableData.subscription_id')));
+    var subscriptions = await models.Subscription.find({ uuid: { $in: subscriptionUuids } }).limit(subscriptionsLimit).lean({ virtuals: true });
+    await applyQueryFieldsToSubscriptions(subscriptions, queryFields.subscription, args, models);
+
+    _.each(subscriptions, (sub)=>{
+      if(_.isUndefined(sub.channelName)){
+        sub.channelName = sub.channel;
+      }
+    });
+    var subscriptionsByUuid = _.keyBy(subscriptions, 'uuid');
+    _.each(resources, (resource)=>{
+      var subId = resource.searchableData.subscription_id;
+      if(!subId){
+        return;
+      }
+      resource.subscription = subscriptionsByUuid[subId] || null;
+      if(resource.subscription) {
+        delete resource.subscription.channel;
+      }
+    });
+
+    // if(_.get(queryFields, 'resources.subscription.channel')){
+    //
+    //   var channelUuids = _.filter(_.uniq(_.map(resources, 'subscription.channel_uuid')));
+    //   var channels = await models.Channel.find({ uuid: { $in: channelUuids } }).lean({ virtuals: true });
+    //   var channelsByUuid = _.keyBy(channels, 'uuid');
+    //   _.each(resources, (resource)=>{
+    //     if(!resource.subscription){
+    //       return;
+    //     }
+    //     resource.subscription.channel = null;
+    //     resource.subscription.channel = channelsByUuid[resource.subscription.channel_uuid] || null;
+    //   });
+    // }
+  }
+};
+
+const applyQueryFieldsToChannels = async(channels, queryFields={}, args, models)=>{ // eslint-disable-line
+  // resolvers/channel.js has something different here. we'll need to look into porting it later
+};
+
+const applyQueryFieldsToSubscriptions = async(subs, queryFields={}, args, models)=>{ // eslint-disable-line
+  var { orgId } = args;
+
+  _.each(subs, (sub)=>{
+    if(_.isUndefined(sub.channelName)){
+      sub.channelName = sub.channel;
+    }
+    delete sub.channel;
+  });
+  var subUuids = _.uniq(_.map(subs, 'uuid'));
+
+  if(queryFields.channel){
+    var channelUuids = _.uniq(_.map(subs, 'channelUuid'));
+    var channels = await models.Channel.find({ uuid: { $in: channelUuids } });
+    await applyQueryFieldsToChannels(channels, queryFields.channel, args, models);
+
+    var channelsByUuid = _.keyBy(channels, 'uuid');
+    _.each(subs, (sub)=>{
+      var channelUuid = sub.channelUuid;
+      var channel = channelsByUuid[channelUuid];
+      sub.channel = channel;
+    });
+  }
+  if(queryFields.resources){
+    var resources = await models.Resource.find({ org_id: orgId, 'searchableData.subscription_id' : { $in: subUuids } }).lean({ virtuals: true });
+    var resourcesBySubUuid = _.groupBy(resources, 'searchableData.subscription_id');
+    _.each(subs, (sub)=>{
+      sub.resources = resourcesBySubUuid[sub.uuid] || [];
+    });
+  }
+  if(queryFields.groupObjs){
+    var groupNames = _.flatten(_.map(subs, 'groups'));
+    var groups = await models.Group.find({ org_id: orgId, name: { $in: groupNames } });
+    await applyQueryFieldsToGroups(groups, queryFields.groupObjs, args, models);
+
+    var groupsByName = _.keyBy(groups, 'name');
+    _.each(subs, (sub)=>{
+      sub.groupObjs = _.filter(_.map(sub.groups, (groupName)=>{
+        return groupsByName[groupName];
+      }));
+    });
+  }
+  if(queryFields.remoteResources || queryFields.rolloutStatus){
+    var remoteResources = await models.Resource.find({
+      org_id: orgId,
+      'searchableData.annotations["deploy_razee_io_clustersubscription"]': { $in: subUuids },
+      deleted: false,
+    });
+    await applyQueryFieldsToResources(resources, queryFields.remoteResources, args, models);
+
+    var remoteResourcesBySubUuid = _.groupBy(remoteResources, (rr)=>{
+      return rr.searchableData.get('annotations["deploy_razee_io_clustersubscription"]');
+    });
+    _.each(subs, (sub)=>{
+      var rrs = remoteResourcesBySubUuid[sub.uuid] || [];
+
+      // loops through each resource. if there are errors, increments the errorCount. if no errors, increments successfulCount
+      var errorCount = 0;
+      var successCount = 0;
+      _.each(rrs, (rr)=>{
+        var errors = _.toArray(rr.searchableData.get('errors')||[]);
+        if(errors.length > 0){
+          errorCount += 1;
+        }
+        else{
+          successCount += 1;
+        }
+      });
+
+      sub.remoteResources = rrs;
+      sub.rolloutStatus = {
+        successCount,
+        errorCount,
+      };
+    });
+  }
+};
+
+module.exports = {
+  applyQueryFieldsToChannels,
+  applyQueryFieldsToClusters,
+  applyQueryFieldsToGroups,
+  applyQueryFieldsToResources,
+  applyQueryFieldsToSubscriptions,
+};
